CitySelect retirado de Adapters
