+++ frontend/src/views/components/Adapters.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react'
 import { TextField, Label, DatePicker } from 'office-ui-fabric-react'
-import Select, { Async } from 'react-select'
+import Select from 'react-select'
 import VirtualizedSelect from 'react-virtualized-select'
 import { Field } from 'react-final-form'
 
-import fetcher from '../../helpers/fetcher'
+import CitySelect from './CitySelect'
 
 export const TextFieldAdapter = ({ input, label, meta, ...rest }) => (
   <TextField
@@ -104,77 +104,4 @@ class BasicSelect extends Component {
       onChange={value => this.onChange(value, this.props.name)}
     />
   }
-}
-
-class CitySelect extends Component {
-  constructor (props) {
-    super(props)
-
-    this.state = {
-      selectedState: null,
-      selectedCity: null,
-      cityList: [],
-      stateList: []
-    }
-
-    this.updateSelectedState = this.updateSelectedState.bind(this)
-    this.selectCity = this.selectCity.bind(this)
-  }
-
-  updateSelectedState (item) {
-    if (!item) {
-      this.setState({ selectedState: null, cityList: [] })
-      return // Fix bug on clear input.
-    }
-
-    fetcher.cachedGet(`/cidade/estado/${item.value}`)
-      .then(response => {
-        const cityList = response.data.map(item => ({ value: item.cidadeId, label: item.nome }))
-        this.setState({ cityList: cityList })
-      })
-
-    this.setState({ selectedState: item.value })
-  }
-
-  selectCity (item) {
-    if (!item) {
-      this.setState({ selectedCity: null })
-      return // Fix bug on clear input.
-    }
-
-    this.setState({ selectedCity: item.value })
-    this.props.onChange(({ cidadeId: item.value }))
-  }
-
-  render () {
-    return <div className='ms-TextField-wrapper'>
-      <Label>{ this.props.label }</Label>
-      <div className='Form-Select-City'>
-        <Async
-          placeholder='UF'
-          searchable
-          style={{width: 100}}
-          value={this.state.selectedState}
-          loadOptions={LoadStateList}
-          onChange={this.updateSelectedState}
-        />
-        <Select
-          placeholder='Cidade'
-          searchable
-          value={this.state.selectedCity}
-          disabled={this.state.cityList.length === 0}
-          options={this.state.cityList}
-          onChange={this.selectCity}
-        />
-      </div>
-    </div>
-  }
-}
-
-const LoadStateList = input => {
-  return fetcher.cachedGet('/estado/all')
-    .then(response => {
-      const selectOptions = response.data.map(item => ({ value: item.estadoId, label: item.uf }))
-      return { options: selectOptions }
-    })
 }
